refactor: move epoch 2.x inventory state machine and tests into `src/net/inv/` module and name their files `epoch2x.rs`
