+++ stackslib/src/net/inv/mod.rs
@@ -1,0 +1,26 @@
+// Copyright (C) 2013-2020 Blockstack PBC, a public benefit corporation
+// Copyright (C) 2020-2023 Stacks Open Internet Foundation
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pub mod epoch2x;
+pub mod nakamoto;
+
+#[cfg(test)]
+pub mod tests;
+
+// Stacks 2.x inventory state machine
+pub use crate::net::inv::epoch2x as inv2x;
+
+pub use inv2x::{INV_REWARD_CYCLES, INV_SYNC_INTERVAL};
